Add Config support to ProducerSettings
