+++ jms/src/main/resources/reference.conf
@@ -7,6 +7,7 @@ alpakka.jms {
     # infinite, or positive integer
     max-retries = 10
   }
+
   send-retry {
     initial-retry = 20 millis
     backoff-factor = 1.5
@@ -16,6 +17,7 @@ alpakka.jms {
 
   consumer {
     connection-retry = ${alpakka.jms.connection-retry}
+    credentials = off
 //    credentials {
 //      username = "some text"
 //      password = "some text"
@@ -29,6 +31,19 @@ alpakka.jms {
     acknowledge-mode = off
     ack-timeout = 1 second
     //durable-name = off // "some text" # optional
+  }
+
+  producer {
+    connection-retry = ${alpakka.jms.connection-retry}
+    send-retry = ${alpakka.jms.send-retry}
+    credentials = off
+//    credentials {
+//      username = "some text"
+//      password = "some text"
+//    }
+    session-count = 1
+    time-to-live = off
+    acknowledge-mode = off
   }
 
   browse {
