+++ jms/src/test/java/docs/javadsl/JmsSettingsTest.java
@@ -24,21 +24,19 @@ import com.typesafe.config.Config;
 public class JmsSettingsTest {
 
   @Test
-  public void consumerSettings() {}
-
-  @Test
   public void settings() throws Exception {
 
     Config config = ConfigFactory.load();
     // #retry-settings
     Config connectionRetryConfig = config.getConfig("alpakka.jms.connection-retry");
+    // reiterating the values form reference.conf
     ConnectionRetrySettings retrySettings =
         ConnectionRetrySettings.create(connectionRetryConfig)
-            .withConnectTimeout(Duration.ofSeconds(3))
-            .withInitialRetry(Duration.ofSeconds(1))
-            .withBackoffFactor(1.5)
-            .withMaxBackoff(Duration.ofSeconds(30))
-            .withInfiniteRetries();
+            .withConnectTimeout(Duration.ofSeconds(10))
+            .withInitialRetry(Duration.ofMillis(100))
+            .withBackoffFactor(2.0)
+            .withMaxBackoff(Duration.ofMinutes(1))
+            .withMaxRetries(10);
     // #retry-settings
 
     ConnectionRetrySettings retrySettings2 = ConnectionRetrySettings.create(connectionRetryConfig);
@@ -48,17 +46,18 @@ public class JmsSettingsTest {
     Config sendRetryConfig = config.getConfig("alpakka.jms.send-retry");
     SendRetrySettings sendRetrySettings =
         SendRetrySettings.create(sendRetryConfig)
-            .withInitialRetry(Duration.ofSeconds(10))
-            .withBackoffFactor(2)
-            .withMaxBackoff(Duration.ofSeconds(1))
-            .withMaxRetries(60);
+            .withInitialRetry(Duration.ofMillis(20))
+            .withBackoffFactor(1.5d)
+            .withMaxBackoff(Duration.ofMillis(500))
+            .withMaxRetries(10);
     // #send-retry-settings
     SendRetrySettings sendRetrySettings2 = SendRetrySettings.create(sendRetryConfig);
     assertEquals(sendRetrySettings.toString(), sendRetrySettings2.toString());
 
     // #producer-settings
+    Config producerConfig = config.getConfig(JmsProducerSettings.configPath());
     JmsProducerSettings settings =
-        JmsProducerSettings.create(new ActiveMQConnectionFactory("broker-url"))
+        JmsProducerSettings.create(producerConfig, new ActiveMQConnectionFactory("broker-url"))
             .withTopic("target-topic")
             .withCredential(Credentials.create("username", "password"))
             .withConnectionRetrySettings(retrySettings)
